requirements.txt
sqlalchemy<2.0.0
werkzeug==2.1.1
fastapi==0.75
pydantic[email,dotenv]==1.9
python-jose==3.3.0
uvicorn==0.17.6
passlib==1.7.4
aiomysql==0.1.0
python-multipart==0.0.5

alembic==1.7.7
